requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
alembic=1.7.7=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
ca-certificates=2022.2.1=haa95532_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cpuonly=2.0=0
cycler=0.11.0=pypi_0
databricks-cli=0.16.4=pypi_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distlib=0.3.4=pypi_0
docker=5.0.3=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py39haa95532_0
flask=2.0.3=pypi_0
fonttools=4.30.0=pypi_0
freetype=2.10.4=hd328e21_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.44.0=pypi_0
icu=58.2=ha925a31_3
idna=3.3=pypi_0
importlib-metadata=4.8.2=py39haa95532_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py39haa95532_1
ipython=7.31.1=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.1.1=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py39haa95532_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lxml=4.8.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
mako=1.2.0=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py39h2bbff1b_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
mlflow=1.24.0=pypi_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.1.0=py39haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.8=py39haa95532_0
numpy=1.22.3=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1m=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=pypi_0
pandas-read-xml=0.3.1=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
pipreqs=0.4.11=pypi_0
prometheus-flask-exporter=0.19.0=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.4=pypi_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.0=py39h196d8e1_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.1=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pytz=2021.3=pypi_0
pywin32=227=pypi_0
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py39hd77b12b_2
qt=5.9.7=vc14h73c81de_0
qtconsole=5.2.2=pyhd3eb1b0_0
qtpy=1.11.2=pyhd3eb1b0_0
querystring-parser=1.2.4=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py39haa95532_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlalchemy=1.4.32=pypi_0
sqlite=3.38.0=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.1=py39haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
thop=0.0.31-2005241907=pypi_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.10.1=py39_cpu
torchvision=0.11.2=py39_cpu
tornado=6.1=py39h2bbff1b_0
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
waitress=2.1.1=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=1.3.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39haa95532_0
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
xmltodict=0.12.0=pypi_0
xz=5.2.5=h62dcd97_0
yarg=0.1.9=pypi_0
zipfile36=0.1.3=pypi_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
